Implemented visually superior update system in `Dashboard` view
The old update system would only update the player layer and the marker layer together.
Since markers need to be fetched first, there might be a long delay between the player moving and it being displayed on the map, as it was dependent solely on the delay.
As Overpass delays are quite disastrous, this update system will immediately update the player and fix issues where positions were drawn for already passed promises.
Also, the update system now performs eased visual transitions that don't look as jaggy on the eye.